Revert "Revert "tiles an fix little""
This reverts commit a0099ac0498c9c13bcab19b7599cdbb3ad614568.

+++ LevelBuildingSidekick/Assets/ISILab/Commons/ConstraintSolver/WFC/WFCSolver.cs
@@ -8,7 +8,7 @@ public class TileConnectWFC
     private int rotation; // max 3 (*90)
 
     public TileConections Tile => tile;
-    public float Rotation => rotation; // el 90 puede ser parametrizado o calculado con el 4
+    public int Rotation => rotation; // el 90 puede ser parametrizado o calculado con el 4
 
     public TileConnectWFC(TileConections tile, int rotation)
     {
